requirements.txt
tensorflow>=2.15,<3.0
numpy>=1.24
Pillow>=10.0
scikit-learn>=1.3
streamlit>=1.35
